添加管理员消息接收功能，修复部分bug

+++ src/main/java/com/cchl/dao/MajorMapper.java
@@ -11,5 +11,5 @@ public interface MajorMapper {
 
     List<Major> selectAll();
 
-    List<Major> selectByDepartmentId(Long departmentId);
+    List<Major> selectByDepartmentId(Integer departmentId);
 }
